Clear the remaining compiler warnings
Two were real defects. The dialog buttons in a private message tab were
wired to setupChannelModeButtons:, which does not exist, so clicking one
did nothing; they call doDialogButton:. And ChatViewController still
forwarded splitViewDidResizeSubviews: to a method the split view
rewrite removed.

The rest were format specifiers needing casts, notifications passed as
nil to parameters marked non-null, one-bit signed bit-fields that cannot
hold 1, undeclared selectors, a mismatched style mask parameter, a
duplicate macro, and two functions in the core that this build compiles
out. First-party code builds clean.

+++ Mac/Sources/SystemVersion.m
@@ -47,7 +47,7 @@ static SystemVersion *SystemVersionSharedSystemVersion;
         self->minor = version.minorVersion;
         self->bugfix = version.patchVersion;
 
-        systemBranch = [[NSString alloc] initWithFormat:@"%d.%d", major, minor];
+        systemBranch = [[NSString alloc] initWithFormat:@"%d.%d", (int)major, (int)minor];
     }
     return self;
 }
